@@ -10,7 +10,15 @@ namespace Drupal\Core\FileTransfer;
  * to the server using some backend (for example FTP or SSH). To keep security,
  * the password should always be asked from the user and never stored. For
  * safety, all methods operate only inside a "jail", by default the Drupal root.
+ *
+ * The following properties are managed by magic methods:
+ *
+ * @property string|false|null $chroot
+ *   Path to connection chroot.
+ * @property object|false|null $connection
+ *   The instantiated connection object.
  */
+#[\AllowDynamicProperties]
 abstract class FileTransfer {
 
   /**
@@ -47,20 +55,6 @@ abstract class FileTransfer {
    * @var string
    */
   protected $jail;
-
-  /**
-   * Path to connection chroot.
-   *
-   * @var string|false|null
-   */
-  private $chrootPath;
-
-  /**
-   * The instantiated connection object.
-   *
-   * @var object|false|null
-   */
-  private $connectionHandle;
 
   /**
    * Constructs a Drupal\Core\FileTransfer\FileTransfer object.
@@ -114,12 +108,12 @@ abstract class FileTransfer {
   public function __get($name) {
     if ($name == 'connection') {
       $this->connect();
-      return $this->connectionHandle;
+      return $this->connection;
     }
 
     if ($name == 'chroot') {
       $this->setChroot();
-      return $this->chrootPath;
+      return $this->chroot;
     }
   }
 
@@ -128,10 +122,10 @@ abstract class FileTransfer {
    */
   public function __set(string $name, $value): void {
     if ($name == 'connection') {
-      $this->connectionHandle = $value;
+      $this->connection = $value;
     }
     elseif ($name == 'chroot') {
-      $this->chrootPath = $value;
+      $this->chroot = $value;
     }
   }
 
@@ -140,10 +134,10 @@ abstract class FileTransfer {
    */
   public function __isset(string $name): bool {
     if ($name == 'connection') {
-      return isset($this->connectionHandle);
+      return isset($this->connection);
     }
     if ($name == 'chroot') {
-      return isset($this->chrootPath);
+      return isset($this->chroot);
     }
     return FALSE;
   }
@@ -153,10 +147,10 @@ abstract class FileTransfer {
    */
   public function __unset(string $name): void {
     if ($name == 'connection') {
-      unset($this->connectionHandle);
+      unset($this->connection);
     }
     elseif ($name == 'chroot') {
-      unset($this->chrootPath);
+      unset($this->chroot);
     }
   }
 
@@ -293,8 +287,8 @@ abstract class FileTransfer {
     // Strip out windows drive letter if its there.
     $path = preg_replace('|^([a-z]{1}):|i', '', $path);
     if ($strip_chroot) {
-      if ($this->chrootPath && str_starts_with($path, $this->chrootPath)) {
-        $path = ($path == $this->chrootPath) ? '' : substr($path, strlen($this->chrootPath));
+      if ($this->chroot && str_starts_with($path, $this->chroot)) {
+        $path = ($path == $this->chroot) ? '' : substr($path, strlen($this->chroot));
       }
     }
     return $path;
