@@ -31,9 +31,9 @@ final class PhpRequirements {
    *   by the PHP version.
    */
   private static $phpEolDates = [
-    '8.1' => '2024-11-25',
-    '8.2' => '2025-12-08',
-    '8.3' => '2026-11-23',
+    '8.1' => '2025-12-31',
+    '8.2' => '2026-12-31',
+    '8.3' => '2027-12-31',
   ];
 
   /**
