@@ -14,8 +14,9 @@
    *   The HTML for the progress bar.
    */
   Drupal.theme.progressBar = function (id) {
+    const escapedId = Drupal.checkPlain(id);
     return (
-      `<div id="${id}" class="progress" aria-live="polite">` +
+      `<div id="${escapedId}" class="progress" aria-live="polite">` +
       '<div class="progress__label">&nbsp;</div>' +
       '<div class="progress__track"><div class="progress__bar"></div></div>' +
       '<div class="progress__percentage"></div>' +
