@@ -336,18 +336,65 @@ class ContactPersonalTest extends BrowserTestBase {
    * @param array $message
    *   (optional) An array with the form fields being used. Defaults to an empty
    *   array.
+   * @param bool $user_copy
+   *   (optional) A boolean to determine whether to send a user copy email.
+   *   Defaults to FALSE.
    *
    * @return array
    *   An array with the form fields being used.
    */
-  protected function submitPersonalContact(AccountInterface $account, array $message = []) {
+  protected function submitPersonalContact(AccountInterface $account, array $message = [], bool $user_copy = FALSE) {
     $message += [
       'subject[0][value]' => $this->randomMachineName(16) . '< " =+ >',
       'message[0][value]' => $this->randomMachineName(64) . '< " =+ >',
+      'copy' => $user_copy,
     ];
     $this->drupalGet('user/' . $account->id() . '/contact');
     $this->submitForm($message, 'Send message');
     return $message;
+  }
+
+  /**
+   * Tests that the opt-out message is included correctly in contact emails.
+   */
+  public function testPersonalContactForm(): void {
+    $opt_out_message = "If you don't want to receive such messages, you can change your settings at";
+
+    // Send an email from an admin (should not contain the opt-out message).
+    $this->drupalLogin($this->adminUser);
+    $this->submitPersonalContact($this->contactUser);
+    $this->drupalLogout();
+
+    $this->assertStringNotContainsString($opt_out_message, $this->getMails()[0]['body'], 'Opt-out message excluded in email.');
+
+    // Send an email from a non-admin (should contain the opt-out message).
+    $this->drupalLogin($this->webUser);
+    $this->submitPersonalContact($this->contactUser);
+
+    $this->assertMailString('body', $opt_out_message, 1, 'Opt-out message included in email.');
+  }
+
+  /**
+   * Tests that the opt-out message is not included in user copy emails.
+   */
+  public function testPersonalContactFormUserCopy(): void {
+    $opt_out_message = "If you don't want to receive such messages, you can change your settings at";
+
+    // Send an email from an admin.
+    $this->drupalLogin($this->adminUser);
+    $this->submitPersonalContact($this->contactUser, [], TRUE);
+    $this->drupalLogout();
+
+    // Send an email from a non-admin.
+    $this->drupalLogin($this->webUser);
+    $this->submitPersonalContact($this->contactUser, [], TRUE);
+
+    $user_copy_emails = $this->getMails(['id' => 'contact_user_copy']);
+
+    // Tests that the opt-out message is not included in admin user copy emails.
+    $this->assertStringNotContainsString($opt_out_message, $user_copy_emails[0]['body'], 'Opt-out message not included in admin user copy email.');
+    // Tests that the opt-out message is not included in non-admin user copy emails.
+    $this->assertStringNotContainsString($opt_out_message, $user_copy_emails[1]['body'], 'Opt-out message not included in non-admin user copy email.');
   }
 
 }
