@@ -2,7 +2,6 @@
 
 namespace Drupal\KernelTests;
 
-use Drupal\Component\Render\FormattableMarkup;
 use Drupal\Core\Form\FormState;
 
 /**
@@ -53,12 +52,9 @@ abstract class ConfigFormTestBase extends KernelTestBase {
     // Check that the form returns an error when expected, and vice versa.
     $errors = $form_state->getErrors();
     $valid_form = empty($errors);
-    $args = [
-      '%values' => print_r($values, TRUE),
-      '%errors' => $valid_form ? t('None') : implode(' ', $errors),
-    ];
-    $this->assertTrue($valid_form, new FormattableMarkup('Input values: %values<br/>Validation handler errors: %errors', $args));
-
+    $values = print_r($values, TRUE);
+    $errors = $valid_form ? t('None') : implode(' ', $errors);
+    $this->assertTrue($valid_form, sprintf('Input values: %s<br/>Validation handler errors: %s', $values, $errors));
     foreach ($this->values as $data) {
       $this->assertEquals($this->config($data['#config_name'])->get($data['#config_key']), $data['#value']);
     }
