@@ -2,8 +2,6 @@
 
 namespace Drupal\KernelTests\Core\Database;
 
-use Drupal\Component\Render\FormattableMarkup;
-
 /**
  * Tests the Insert query builder with LOB fields.
  *
@@ -21,7 +19,7 @@ class InsertLobTest extends DatabaseTestBase {
       ->fields(['blob1' => $data])
       ->execute();
     $r = $this->connection->query('SELECT * FROM {test_one_blob} WHERE [id] = :id', [':id' => $id])->fetchAssoc();
-    $this->assertSame($data, $r['blob1'], new FormattableMarkup('Can insert a blob: id @id, @data.', ['@id' => $id, '@data' => serialize($r)]));
+    $this->assertSame($data, $r['blob1'], "Can insert a blob: id $id, " . serialize($r));
   }
 
   /**
