@@ -2,8 +2,6 @@
 
 namespace Drupal\KernelTests\Core\Database;
 
-use Drupal\Component\Render\FormattableMarkup;
-
 /**
  * Tests the Update query builder with LOB fields.
  *
@@ -28,7 +26,7 @@ class UpdateLobTest extends DatabaseTestBase {
       ->execute();
 
     $r = $this->connection->query('SELECT * FROM {test_one_blob} WHERE [id] = :id', [':id' => $id])->fetchAssoc();
-    $this->assertSame($data, $r['blob1'], new FormattableMarkup('Can update a blob: id @id, @data.', ['@id' => $id, '@data' => serialize($r)]));
+    $this->assertSame($data, $r['blob1'], "Can update a blob: id $id, " . serialize($r));
   }
 
   /**
