@@ -8,6 +8,7 @@ use Drupal\Core\Cache\NullBackend;
 use Drupal\Core\DependencyInjection\ContainerBuilder;
 use Drupal\Core\TypedData\DataDefinition;
 use Drupal\Core\TypedData\MapDataDefinition;
+use Drupal\Core\TypedData\TypedData as TypedDataBase;
 use Drupal\Core\TypedData\TypedDataManager;
 use Drupal\Core\TypedData\Validation\ExecutionContextFactory;
 use Drupal\Core\TypedData\Validation\RecursiveValidator;
@@ -250,16 +251,13 @@ class RecursiveContextualValidatorTest extends UnitTestCase {
 
   /**
    * Provides data for testValidatePropertyWithInvalidObjects.
-   * @return array
    */
-  public function providerTestValidatePropertyWithInvalidObjects() {
-    $data = [];
-    $data[] = [new \stdClass()];
-    $data[] = [new class() {}];
+  public static function providerTestValidatePropertyWithInvalidObjects(): \Generator {
+    $dataDefinition = new DataDefinition();
 
-    $data[] = [$this->createMock('Drupal\Core\TypedData\TypedDataInterface')];
-
-    return $data;
+    yield [new \stdClass()];
+    yield [new class() {}];
+    yield [new class($dataDefinition) extends TypedDataBase {}];
   }
 
   /**
